Add Maths, Physics, Time, and Collision modules. Implement movement of paddle and ball. Implement collision algorithm (Swept AABB).

+++ lib/canvas.ml
@@ -1,5 +1,5 @@
 module Canvas = struct
-    let window_width : int = 640
-    let window_height : int = 480
-    let margin_size : int = 5
+    let window_width : float = 640.0
+    let window_height : float = 480.0
+    let margin_size : float = 5.0
 end
